+++ tie_break/tests/integration_test.rs
@@ -27,6 +27,17 @@ const TEST_SETS: &[TestSet] = &[
         expected_summary_file: "test_data/tb_small/alignment_summary.csv",
     },
     TestSet {
+        name: "no_mapping",
+        input_bam: "test_data/no_mapping/aln.bam",
+        contigs: "test_data/no_mapping/contigs.csv",
+        output_root: "tests/test_produced_files/no_mapping/",
+        parameters: "test_data/no_mapping/params.yaml",
+        depth_counts_file: "test_data/no_mapping/depth_counts.csv",
+        expected_ref_tie_breaker_order_file: "test_data/no_mapping/ref_tie_breaker_order.csv",
+        expected_stats_file: "test_data/no_mapping/stats.yaml",
+        expected_summary_file: "test_data/no_mapping/alignment_summary.csv",
+    },
+    TestSet {
         name: "tb_intracellulare",
         input_bam: "test_data/tb_intracellulare/aln.bam",
         contigs: "test_data/tb_intracellulare/contigs.csv",
@@ -179,7 +190,11 @@ fn test_debug() {
     );
 
     assert!(
-        round1_depth_equal && ref_tie_breaker_order_equal && stats_equal && depths_equal && debug_equal,
+        round1_depth_equal
+            && ref_tie_breaker_order_equal
+            && stats_equal
+            && depths_equal
+            && debug_equal,
         "Test set '{}' failed",
         test_set.name
     );