+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## New
+
+- Refactor tie_break to use signal in alns csv internally
+- Fixed bug where winner read count didn't include unique reads (but should have done)
+- add "good" rows to alignment summary. This includes shared reads from round 1
+- tie_break produces sensible output on empty bam
+
 ## 3.0.1
 
 - Correct file extension for build system